Chore: fix circular dependency
