+++ public/app/features/explore/TraceView/createTraceLink.ts
@@ -4,6 +4,7 @@ import {
   type DataLinkPostProcessor,
   type DataSourceInstanceSettings,
   type DataSourceJsonData,
+  dateTime,
   type Field,
   FieldType,
   type LinkModel,
@@ -22,7 +23,6 @@ import { getVariableUsageInfo } from '../utils/links';
 import { getLogsButtonCTA } from './components/TraceTimelineViewer/SpanDetail/LogsLink';
 import { getTraceToLogsTraceQuery, interpolateQueries } from './components/logsLink';
 import { type Trace } from './components/types/trace';
-import { getTimeRangeFromTimestamps } from './createSpanLink';
 
 /**
  * Builds an Explore link model for related logs at the trace level (no span id filter).
@@ -146,6 +146,43 @@ function getTimeRangeFromTrace(
   isSplunkDS = false
 ): TimeRange {
   return getTimeRangeFromTimestamps(trace.startTime, trace.duration, timeShift, isSplunkDS);
+}
+
+export function getTimeRangeFromTimestamps(
+  startTimeUs: number,
+  durationUs: number,
+  timeShift: { startMs: number; endMs: number } = { startMs: 0, endMs: 0 },
+  isSplunkDS = false,
+  shouldCreatePyroscopeLink = false
+): TimeRange {
+  let adjustedStartTime = Math.floor(startTimeUs / 1000 + timeShift.startMs);
+  const endMs = (startTimeUs + durationUs) / 1000;
+  let adjustedEndTime = Math.floor(endMs + timeShift.endMs);
+
+  // Splunk requires a time interval of >= 1s, rather than >=1ms like Loki timerange in below elseif block
+  if (isSplunkDS && adjustedEndTime - adjustedStartTime < 1000) {
+    adjustedEndTime = adjustedStartTime + 1000;
+  } else if (shouldCreatePyroscopeLink) {
+    adjustedStartTime = adjustedStartTime - 60000;
+    adjustedEndTime = adjustedEndTime + 60000;
+  } else if (adjustedStartTime >= adjustedEndTime) {
+    // Because we can only pass milliseconds in the url we need to check if they are greater or equal.
+    // We need end time to be later than start time
+    adjustedEndTime = adjustedStartTime + 1;
+  }
+
+  const to = dateTime(adjustedEndTime);
+  const from = dateTime(adjustedStartTime);
+
+  // Beware that public/app/features/explore/state/main.ts SplitOpen fn uses the range from here. No matter what is in the url.
+  return {
+    from,
+    to,
+    raw: {
+      from,
+      to,
+    },
+  };
 }
 
 /**
